+++ src/db.rs
@@ -1,15 +1,13 @@
-use crate::config::{Config, DatabaseConfig};
+use crate::config::Config;
 use crate::web::auth::{TwitchUserAccessToken, UserAuthorization};
 use chrono::{DateTime, Utc};
 use deadpool_postgres::{ManagerConfig, PoolConfig, RecyclingMethod};
 use itertools::Itertools;
-use murmur3::murmur3_32;
-use prometheus::{HistogramVec, IntCounterVec, IntGaugeVec};
-use prometheus::{register_histogram_vec, register_int_counter_vec, register_int_gauge_vec};
+use prometheus::{
+    Histogram, IntCounter, IntGauge, register_histogram, register_int_counter, register_int_gauge,
+};
 use rustls::ClientConfig;
 use rustls_platform_verifier::ConfigVerifierExt;
-use std::fmt::{Display, Formatter};
-use std::io::Cursor;
 use std::ops::DerefMut;
 use std::time::Duration;
 use std::{collections::HashSet, sync::LazyLock};
@@ -18,155 +16,71 @@ use tokio_postgres::types::ToSql;
 use tokio_postgres_rustls::MakeRustlsConnect;
 use tokio_util::sync::CancellationToken;
 
-static MESSAGES_APPENDED: LazyLock<IntCounterVec> = LazyLock::new(|| {
-    register_int_counter_vec!(
+static MESSAGES_APPENDED: LazyLock<IntCounter> = LazyLock::new(|| {
+    register_int_counter!(
         "recentmessages_messages_appended",
-        "Total number of messages appended to storage",
-        &["db"]
+        "Total number of messages appended to storage"
     )
     .unwrap()
 });
 
-static MESSAGES_STORED: LazyLock<IntGaugeVec> = LazyLock::new(|| {
-    register_int_gauge_vec!(
+static MESSAGES_STORED: LazyLock<IntGauge> = LazyLock::new(|| {
+    register_int_gauge!(
         "recentmessages_messages_stored",
-        "Number of messages currently stored in storage",
-        &["db"]
+        "Number of messages currently stored in storage"
     )
     .unwrap()
 });
-static STORE_CHUNK_RUNS: LazyLock<IntCounterVec> = LazyLock::new(|| {
-    register_int_counter_vec!(
-        "recentmessages_irc_forwarder_store_chunk_runs",
-        "Number of runs the IRC forwarder has completed",
-        &["db"]
-    )
-    .unwrap()
-});
-static STORE_CHUNK_ERRORS: LazyLock<IntCounterVec> = LazyLock::new(|| {
-    register_int_counter_vec!(
-        "recentmessages_irc_forwarder_store_chunk_errors",
-        "Number of times a chunk could not be appended to the database successfully",
-        &["db"]
-    )
-    .unwrap()
-});
-static STORE_CHUNK_TIME_TAKEN: LazyLock<HistogramVec> = LazyLock::new(|| {
-    register_histogram_vec!(
-        "recentmessages_irc_forwarder_store_chunk_time_taken_seconds",
-        "Time taken to forward individual chunks of messages to the database",
-        &["db"]
-    )
-    .unwrap()
-});
-static MESSAGES_VACUUMED: LazyLock<IntCounterVec> = LazyLock::new(|| {
-    register_int_counter_vec!(
+static MESSAGES_VACUUMED: LazyLock<IntCounter> = LazyLock::new(|| {
+    register_int_counter!(
         "recentmessages_messages_vacuumed",
-        "Total number of messages that were removed by the automatic vacuum runner",
-        &["db"]
+        "Total number of messages that were removed by the automatic vacuum runner"
     )
     .unwrap()
 });
-static VACUUM_RUNS: LazyLock<IntCounterVec> = LazyLock::new(|| {
-    register_int_counter_vec!(
+static VACUUM_RUNS: LazyLock<IntCounter> = LazyLock::new(|| {
+    register_int_counter!(
         "recentmessages_message_vacuum_runs",
-        "Total number of times the automatic vacuum runner has been started for a certain channel",
-        &["db"]
+        "Total number of times the automatic vacuum runner has been started for a certain channel"
     )
     .unwrap()
 });
-static DB_CONNECTIONS_IN_USE: LazyLock<IntGaugeVec> = LazyLock::new(|| {
-    register_int_gauge_vec!(
+static DB_CONNECTIONS_IN_USE: LazyLock<IntGauge> = LazyLock::new(|| {
+    register_int_gauge!(
         "recentmessages_db_pool_connections_in_use",
-        "Number of database connections currently in use",
-        &["db"]
+        "Number of database connections currently in use"
     )
     .unwrap()
 });
-static DB_CONNECTIONS_MAX: LazyLock<IntGaugeVec> = LazyLock::new(|| {
-    register_int_gauge_vec!(
+static DB_CONNECTIONS_MAX: LazyLock<IntGauge> = LazyLock::new(|| {
+    register_int_gauge!(
         "recentmessages_db_pool_connections_max",
-        "Configured maximum size of the database connection pool",
-        &["db"]
+        "Configured maximum size of the database connection pool"
     )
     .unwrap()
 });
-static TIME_TAKEN_TO_GET_DB_CONN: LazyLock<HistogramVec> = LazyLock::new(|| {
-    register_histogram_vec!(
+static TIME_TAKEN_TO_GET_DB_CONN: LazyLock<Histogram> = LazyLock::new(|| {
+    register_histogram!(
         "recentmessages_db_pool_retrieval_time_seconds",
-        "Time taken to retrieve a DB connection from the database pool",
-        &["db"]
+        "Time taken to retrieve a DB connection from the database pool"
     )
     .unwrap()
 });
-
-#[derive(Clone)]
-pub struct DatabaseAccess {
-    db_pool: deadpool_postgres::Pool,
-    cached_name: &'static str,
-}
-
-impl DatabaseAccess {
-    /// Warning: this leaks a small amount of memory for the name, but it shouldn't be a problem
-    /// since this happens only once during application startup and the "leaked" value
-    /// is needed for the entirety of the program runtime
-    pub fn new(
-        custom_name: Option<&str>,
-        partition_id: usize,
-        db_pool: deadpool_postgres::Pool,
-    ) -> Self {
-        let shard_or_main = if partition_id == 0 { "main" } else { "shard" };
-        let cached_name = if let Some(custom_name) = &custom_name {
-            format!("db{partition_id}({shard_or_main}, {custom_name})")
-        } else {
-            format!("db{partition_id}({shard_or_main})")
-        };
-        let cached_name = Box::leak(Box::new(cached_name));
-        DatabaseAccess {
-            db_pool,
-            cached_name,
-        }
-    }
-}
-
-impl Display for DatabaseAccess {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.cached_name)
-    }
-}
 
 pub fn connect_to_postgresql(config: &Config) -> DataStorage {
-    let mut partition_id_counter = 0usize;
-    let main_db = connect_to_single_postgres_server(&config.main_db, &mut partition_id_counter);
-    let mut shard_dbs = Vec::new();
-    for shard_db_config in &config.shard_db {
-        shard_dbs.push(connect_to_single_postgres_server(
-            shard_db_config,
-            &mut partition_id_counter,
-        ));
-    }
-
-    DataStorage::new(main_db, shard_dbs)
-}
-
-fn connect_to_single_postgres_server(
-    config: &DatabaseConfig,
-    partition_id_counter: &mut usize,
-) -> DatabaseAccess {
-    let partition_id = *partition_id_counter;
-    *partition_id_counter += 1;
-
-    let pg_config = tokio_postgres::Config::from(config.clone());
-    tracing::debug!("PostgreSQL config for db{}: {:#?}", partition_id, pg_config);
+    let pg_config = tokio_postgres::Config::from(config.db.clone());
+    tracing::debug!("PostgreSQL config: {:#?}", pg_config);
 
     let mgr_config = ManagerConfig {
         recycling_method: RecyclingMethod::Fast,
     };
     let pool_config = PoolConfig {
-        max_size: config.pool.max_size,
-        timeouts: deadpool_postgres::Timeouts::from(config.pool),
+        max_size: config.db.pool.max_size,
+        timeouts: deadpool_postgres::Timeouts::from(config.db.pool),
         ..Default::default()
     };
+    DB_CONNECTIONS_MAX.set(config.db.pool.max_size as i64);
+    DB_CONNECTIONS_IN_USE.set(0);
 
     let tls_config = ClientConfig::with_platform_verifier().unwrap();
     let tls = MakeRustlsConnect::new(tls_config);
@@ -178,27 +92,13 @@ fn connect_to_single_postgres_server(
         .build()
         .unwrap();
 
-    let db = DatabaseAccess::new(config.name.as_deref(), partition_id, db_pool);
-
-    DB_CONNECTIONS_MAX
-        .with_label_values(&[db.cached_name])
-        .set(config.pool.max_size as i64);
-    DB_CONNECTIONS_IN_USE
-        .with_label_values(&[db.cached_name])
-        .set(0);
-
-    db
+    DataStorage { db: db_pool }
 }
 
-mod migrations_main {
+mod migrations {
     use refinery::embed_migrations;
-    // refers to the "migrations_main" directory in the project root
-    embed_migrations!("migrations_main");
-}
-mod migrations_shard {
-    use refinery::embed_migrations;
-    // refers to the "migrations_shard" directory in the project root
-    embed_migrations!("migrations_shard");
+    // refers to the "migrations" directory in the project root
+    embed_migrations!("migrations");
 }
 
 pub type StorageError = deadpool_postgres::PoolError;
@@ -211,93 +111,48 @@ pub struct StoredMessage {
 
 #[derive(Clone)]
 pub struct DataStorage {
-    main_db: DatabaseAccess,
-    shard_dbs: Vec<DatabaseAccess>,
+    db: deadpool_postgres::Pool,
 }
 
-struct WrappedDbConn(deadpool_postgres::Object, &'static str);
+struct WrappedDbConn(deadpool_postgres::Object);
 
 impl WrappedDbConn {
-    pub fn new(inner: deadpool_postgres::Object, db_partition_name: &'static str) -> WrappedDbConn {
-        DB_CONNECTIONS_IN_USE
-            .with_label_values(&[db_partition_name])
-            .inc();
-        WrappedDbConn(inner, db_partition_name)
+    pub fn new(inner: deadpool_postgres::Object) -> WrappedDbConn {
+        DB_CONNECTIONS_IN_USE.inc();
+        WrappedDbConn(inner)
     }
 }
 
 impl Drop for WrappedDbConn {
     fn drop(&mut self) {
-        DB_CONNECTIONS_IN_USE.with_label_values(&[self.1]).dec();
+        DB_CONNECTIONS_IN_USE.dec();
     }
 }
 
 impl DataStorage {
-    pub fn new(main_db: DatabaseAccess, shard_dbs: Vec<DatabaseAccess>) -> DataStorage {
-        DataStorage { main_db, shard_dbs }
-    }
-
-    fn get_partition(&self, partition_id: usize) -> &DatabaseAccess {
-        if partition_id == 0 {
-            &self.main_db
-        } else {
-            // will panic if partition_id is out of bounds
-            self.shard_dbs.get(partition_id - 1).unwrap()
-        }
-    }
-
-    async fn get_db_conn(&self, partition_id: usize) -> Result<WrappedDbConn, StorageError> {
-        let timer = TIME_TAKEN_TO_GET_DB_CONN
-            .with_label_values(&[self.name_partition(partition_id)])
-            .start_timer();
-        let db_conn = self.get_partition(partition_id).db_pool.get().await;
+    async fn get_db_conn(&self) -> Result<WrappedDbConn, StorageError> {
+        let timer = TIME_TAKEN_TO_GET_DB_CONN.start_timer();
+        let db_conn = self.db.get().await;
         timer.observe_duration();
-        Ok(WrappedDbConn::new(
-            db_conn?,
-            self.name_partition(partition_id),
-        ))
-    }
-
-    async fn get_db_conn_main(&self) -> Result<WrappedDbConn, StorageError> {
-        self.get_db_conn(0).await
-    }
-
-    fn name_partition(&self, partition_id: usize) -> &'static str {
-        self.get_partition(partition_id).cached_name
-    }
-
-    fn channel_to_partition_id(&self, channel_login: &str) -> usize {
-        let hash_result: u32 = murmur3_32(&mut Cursor::new(channel_login), 0).unwrap();
-        (hash_result % ((self.shard_dbs.len() + 1) as u32)) as usize
+        Ok(WrappedDbConn::new(db_conn?))
     }
 
     pub async fn run_migrations(&self) -> Result<(), Box<dyn std::error::Error>> {
-        migrations_main::migrations::runner()
-            .run_async(self.get_db_conn_main().await?.0.as_mut().deref_mut())
+        migrations::migrations::runner()
+            .run_async(self.get_db_conn().await?.0.as_mut().deref_mut())
             .await?;
-
-        for i in 0..self.shard_dbs.len() {
-            migrations_shard::migrations::runner()
-                .run_async(self.get_db_conn(i + 1).await?.0.as_mut().deref_mut())
-                .await?;
-        }
-
         Ok(())
     }
 
     pub async fn fetch_initial_metrics_values(&self) -> Result<(), StorageError> {
-        for i in 0..=self.shard_dbs.len() {
-            let count: i64 = self
-                .get_db_conn(i)
-                .await?
-                .0
-                .query_one("SELECT COUNT(*) AS count FROM message", &[])
-                .await?
-                .get("count");
-            MESSAGES_STORED
-                .with_label_values(&[self.name_partition(i)])
-                .set(count);
-        }
+        let count: i64 = self
+            .get_db_conn()
+            .await?
+            .0
+            .query_one("SELECT COUNT(*) AS count FROM message", &[])
+            .await?
+            .get("count");
+        MESSAGES_STORED.set(count);
         Ok(())
     }
 
@@ -305,7 +160,7 @@ impl DataStorage {
         &self,
         channel_expiry: Duration,
     ) -> Result<HashSet<String>, StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
 
         // TODO figure out whether this has to be sped up using an index.
         let rows = db_conn
@@ -325,7 +180,7 @@ ORDER BY last_access DESC",
     }
 
     pub async fn touch_or_add_channel(&self, channel_login: &str) -> Result<(), StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
         // this way we only update the last_access if it's been at least 30 minutes since
         // the last time the last_access was updated for that channel. For high traffic
         // channels this massively cuts down on the amount of writes the DB has to do
@@ -343,7 +198,7 @@ ON CONFLICT ON CONSTRAINT channel_pkey DO UPDATE
     }
 
     pub async fn is_channel_ignored(&self, channel_login: &str) -> Result<bool, StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
         let rows = db_conn
             .0
             .query(
@@ -362,7 +217,7 @@ WHERE channel_login = $1",
         channel_login: &str,
         ignored: bool,
     ) -> Result<(), StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
         db_conn
             .0
             .query(
@@ -380,7 +235,7 @@ ON CONFLICT ON CONSTRAINT channel_pkey DO UPDATE
         &self,
         user_authorization: &UserAuthorization,
     ) -> Result<(), StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
 
         db_conn
             .0
@@ -410,7 +265,7 @@ VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)",
         &self,
         access_token: &str,
     ) -> Result<Option<UserAuthorization>, StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
 
         let rows = db_conn
             .0
@@ -450,7 +305,7 @@ AND valid_until >= now()",
         &self,
         user_authorization: &UserAuthorization,
     ) -> Result<(), StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
 
         db_conn
             .0
@@ -485,7 +340,7 @@ WHERE access_token = $1",
     // TODO background task to purge expired authorizations
 
     pub async fn delete_user_authorization(&self, access_token: &str) -> Result<(), StorageError> {
-        let db_conn = self.get_db_conn_main().await?;
+        let db_conn = self.get_db_conn().await?;
 
         db_conn
             .0
@@ -508,8 +363,7 @@ WHERE access_token = $1",
         max_buffer_size: usize,
     ) -> Result<Vec<StoredMessage>, StorageError> {
         // limit: If specified, take the newest N messages.
-        let partition_id = self.channel_to_partition_id(channel_login);
-        let db_conn = self.get_db_conn(partition_id).await?;
+        let db_conn = self.get_db_conn().await?;
 
         let limit = match limit {
             Some(limit) => usize::min(limit, max_buffer_size),
@@ -541,9 +395,8 @@ WHERE access_token = $1",
     }
 
     pub async fn purge_messages(&self, channel_login: &str) -> Result<(), StorageError> {
-        let partition_id = self.channel_to_partition_id(channel_login);
         let num_messages_deleted = self
-            .get_db_conn(partition_id)
+            .get_db_conn()
             .await?
             .0
             .execute(
@@ -551,79 +404,35 @@ WHERE access_token = $1",
                 &[&channel_login],
             )
             .await?;
-        MESSAGES_STORED
-            .with_label_values(&[self.name_partition(partition_id)])
-            .sub(num_messages_deleted as i64);
+        MESSAGES_STORED.sub(num_messages_deleted as i64);
         Ok(())
     }
 
     /// Append a message to the storage.
-    pub fn append_messages(&self, messages: Vec<(String, DateTime<Utc>, String)>) {
-        let group_map = messages
-            .into_iter()
-            .into_group_map_by(|(channel_login, _, _)| self.channel_to_partition_id(channel_login));
-
-        for (partition_id, messages) in group_map {
-            let self_clone = self.clone();
-            tokio::spawn(async move {
-                STORE_CHUNK_RUNS
-                    .with_label_values(&[self_clone.name_partition(partition_id)])
-                    .inc();
-                let timer = STORE_CHUNK_TIME_TAKEN
-                    .with_label_values(&[self_clone.name_partition(partition_id)])
-                    .start_timer();
-
-                let res = self_clone
-                    .append_messages_partition(partition_id, messages)
-                    .await;
-                if let Err(e) = res {
-                    tracing::error!(
-                        "Failed to append message chunk to {}: {}",
-                        self_clone.name_partition(partition_id),
-                        e
-                    );
-                    STORE_CHUNK_ERRORS
-                        .with_label_values(&[self_clone.name_partition(partition_id)])
-                        .inc();
-                }
-
-                timer.observe_duration();
-            });
-        }
-    }
-
-    async fn append_messages_partition(
+    pub async fn append_messages(
         &self,
-        partition_id: usize,
-        messages: Vec<(String, DateTime<Utc>, String)>,
+        messages: &[(String, DateTime<Utc>, String)],
     ) -> Result<(), StorageError> {
-        STORE_CHUNK_RUNS
-            .with_label_values(&[self.name_partition(partition_id)])
-            .inc();
-
         if messages.is_empty() {
             return Ok(());
         }
         let num_messages = messages.len();
-        self.get_db_conn(partition_id)
+        self.get_db_conn()
             .await?
             .0
             .execute(
-                &DataStorage::batch_message_insert_query(messages.len(), 3),
-                DataStorage::batch_message_insert_values(&messages).as_slice(),
+                &DataStorage::batch_message_insert_query(num_messages, 3),
+                DataStorage::batch_message_insert_values(messages).as_slice(),
             )
             .await?;
-        MESSAGES_APPENDED
-            .with_label_values(&[self.name_partition(partition_id)])
-            .inc_by(num_messages as u64);
-        MESSAGES_STORED
-            .with_label_values(&[self.name_partition(partition_id)])
-            .add(num_messages as i64);
+        MESSAGES_APPENDED.inc_by(num_messages as u64);
+        MESSAGES_STORED.add(num_messages as i64);
+
         Ok(())
     }
 
     fn batch_message_insert_values(
-        rows: &Vec<(String, DateTime<Utc>, String)>,
+        rows: &[(String, DateTime<Utc>, String)],
     ) -> Vec<&(dyn ToSql + Sync)> {
         let mut out: Vec<&(dyn ToSql + Sync)> = vec![];
         for (a, b, c) in rows {
@@ -670,25 +479,19 @@ WHERE access_token = $1",
             loop {
                 check_interval.tick().await;
                 tracing::info!("Running vacuum for old messages");
-                for partition_id in 0..=self.shard_dbs.len() {
-                    tokio::spawn(async move {
-                        let res = self
-                            .run_message_vacuum(
-                                partition_id,
-                                vacuum_messages_every,
-                                message_expire_after,
-                                max_buffer_size,
-                            )
-                            .await;
+                let res = self
+                    .run_message_vacuum(
+                        vacuum_messages_every,
+                        message_expire_after,
+                        max_buffer_size,
+                    )
+                    .await;
 
-                        if let Err(e) = res {
-                            tracing::error!(
-                                "Failed to start message vacuum batch ({}), skipping entire batch: {}",
-                                self.name_partition(partition_id),
-                                e
-                            );
-                        }
-                    });
+                if let Err(e) = res {
+                    tracing::error!(
+                        "Failed to start message vacuum batch, skipping entire batch: {}",
+                        e
+                    );
                 }
             }
         };
@@ -703,12 +506,11 @@ WHERE access_token = $1",
     /// maximum buffer size.
     async fn run_message_vacuum(
         &self,
-        partition_id: usize,
         vacuum_messages_every: Duration,
         messages_expire_after: Duration,
         max_buffer_size: usize,
     ) -> Result<(), StorageError> {
-        let db_conn = self.get_db_conn(partition_id).await?;
+        let db_conn = self.get_db_conn().await?;
 
         let channels_with_messages: Vec<String> = db_conn
             .0
@@ -727,9 +529,7 @@ WHERE access_token = $1",
 
         for channel in channels_with_messages {
             interval.tick().await;
-            VACUUM_RUNS
-                .with_label_values(&[self.name_partition(partition_id)])
-                .inc();
+            VACUUM_RUNS.inc();
 
             let execute_result = db_conn
                 .0
@@ -761,22 +561,13 @@ AND (
             let messages_deleted = match execute_result {
                 Ok(messages_deleted) => messages_deleted,
                 Err(e) => {
-                    tracing::error!(
-                        "({}) Failed to vacuum channel {}: {}",
-                        self.name_partition(partition_id),
-                        channel,
-                        e
-                    );
+                    tracing::error!("Failed to vacuum channel {}: {}", channel, e);
                     continue;
                 }
             };
 
-            MESSAGES_VACUUMED
-                .with_label_values(&[self.name_partition(partition_id)])
-                .inc_by(messages_deleted);
-            MESSAGES_STORED
-                .with_label_values(&[self.name_partition(partition_id)])
-                .sub(messages_deleted as i64);
+            MESSAGES_VACUUMED.inc_by(messages_deleted);
+            MESSAGES_STORED.sub(messages_deleted as i64);
         }
 
         Ok(())
@@ -787,7 +578,6 @@ AND (
 pub mod test {
     #[test]
     pub fn dump_migrations() {
-        dbg!(super::migrations_main::migrations::runner().get_migrations());
-        dbg!(super::migrations_shard::migrations::runner().get_migrations());
+        dbg!(super::migrations::migrations::runner().get_migrations());
     }
 }